+++ spec/03-plugins/24-rate-limiting/04-access_spec.lua
@@ -174,6 +174,26 @@ for i, policy in ipairs({"local", "cluster", "redis"}) do
         }
       })
 
+      local api5 = assert(helpers.dao.apis:insert {
+        name = "api-5",
+        hosts = { "test5.com" },
+        upstream_url = "http://mockbin.com"
+      })
+      assert(helpers.dao.plugins:insert {
+        name = "rate-limiting",
+        api_id = api5.id,
+        config = {
+          policy = policy,
+          minute = 6,
+          hide_client_headers = true,
+          fault_tolerant = false,
+          redis_host = REDIS_HOST,
+          redis_port = REDIS_PORT,
+          redis_password = REDIS_PASSWORD,
+          redis_database = REDIS_DATABASE
+        }
+      })
+
       assert(helpers.start_kong())
     end)
 
@@ -356,6 +376,22 @@ for i, policy in ipairs({"local", "cluster", "redis"}) do
           local body = assert.res_status(429, res)
           assert.are.equal([[{"message":"API rate limit exceeded"}]], body)
         end)
+      end)
+    end)
+
+    describe("Config with hide_client_headers", function()
+      it("does not send rate-limit headers when hide_client_headers==true", function()
+        local res = assert(helpers.proxy_client():send {
+          method = "GET",
+          path = "/status/200/",
+          headers = {
+            ["Host"] = "test5.com"
+          }
+        })
+
+        assert.res_status(200, res)
+        assert.is_nil(res.headers["x-ratelimit-limit-minute"])
+        assert.is_nil(res.headers["x-ratelimit-remaining-minute"])
       end)
     end)